BalancerIntTest...
