+++ test/baseUT/GasLimits.ts
@@ -32,3 +32,8 @@ export const GAS_CONVERTER_STRATEGY_BASE_CONVERT_WITHDRAW_ALL = 390_000;
 export const GAS_CONVERTER_STRATEGY_BASE_CONVERT_WITHDRAW_EMERGENCY = 322_000;
 
 export const GAS_OPEN_POSITION = 198_000;
+
+export const GAS_DEPOSIT_SIGNER = 5_192_000;
+export const GAS_WITHDRAW_ALL_TO_SPLITTER = 60_001;
+export const GAS_FIRST_HARDWORK = 234_000;
+export const GAS_EMERGENCY_EXIT = 1;
