+++ test/baseUT/utils/UniversalTestUtils.ts
@@ -1,7 +1,7 @@
-import {BigNumber, ContractTransaction} from "ethers";
+import {BigNumber, ContractReceipt, ContractTransaction} from "ethers";
 import hre from "hardhat";
 import {
-  BalancerComposableStableStrategy__factory,
+  BalancerComposableStableStrategy__factory, ControllerV2__factory,
   IBalancerGauge__factory, IController__factory,
   IERC20__factory,
   ISplitter__factory, IStrategyV2,
@@ -16,56 +16,6 @@ import {DeployerUtilsLocal, IVaultStrategyInfo} from "../../../scripts/utils/Dep
 import {DeployerUtils} from "../../../scripts/utils/DeployerUtils";
 import {Misc} from "../../../scripts/utils/Misc";
 import {TokenUtils} from "../../../scripts/utils/TokenUtils";
-import {VaultUtils} from "../../VaultUtils";
-
-/**
- * All balances
- */
-export interface IState {
-  title: string;
-  block: number;
-  blockTimestamp: number;
-  signer: {
-    usdc: BigNumber;
-  }
-  user: {
-    usdc: BigNumber;
-  }
-  strategy: {
-    usdc: BigNumber;
-    usdt: BigNumber;
-    dai: BigNumber;
-    bal: BigNumber;
-    bptPool: BigNumber;
-    totalAssets: BigNumber;
-    investedAssets: BigNumber;
-  }
-  gauge: {
-    strategyBalance: BigNumber;
-  }
-  balancerPool: {
-    bbAmUsdc: BigNumber;
-    bbAmUsdt: BigNumber;
-    bbAmDai: BigNumber;
-  }
-  splitter: {
-    usdc: BigNumber;
-    totalAssets: BigNumber;
-  }
-  vault: {
-    usdc: BigNumber;
-    userShares: BigNumber;
-    signerShares: BigNumber;
-    userUsdc: BigNumber;
-    signerUsdc: BigNumber;
-    sharePrice: BigNumber;
-    totalSupply: BigNumber;
-    totalAssets: BigNumber;
-  },
-  insurance: {
-    usdc: BigNumber;
-  }
-}
 
 export interface IMakeStrategyDeployerInputParams {
   vaultName?: string;
@@ -76,255 +26,21 @@ export interface IMakeStrategyDeployerInputParams {
   wait?: boolean;
 }
 
+export interface IDistributedInfo {
+  sender: string;
+  incomeToken: string;
+  queuedBalance: BigNumber;
+  tetuValue: BigNumber;
+  tetuBalance: BigNumber;
+  toInvestFund: BigNumber;
+  toGauges: BigNumber;
+  toBribes: BigNumber;
+}
+
 /**
  * Utils for universal test
  */
 export class UniversalTestUtils {
-  static async getState(
-    signer: SignerWithAddress,
-    user: SignerWithAddress,
-    strategy: IStrategyV2,
-    vault: TetuVaultV2,
-    title?: string,
-  ) : Promise<IState>{
-    const gauge = "0x1c514fEc643AdD86aeF0ef14F4add28cC3425306";
-    const balancerPool = "0x48e6b98ef6329f8f0a30ebb8c7c960330d648085";
-    const bbAmDai = "0x178E029173417b1F9C8bC16DCeC6f697bC323746";
-    const bbAmUsdc = "0xF93579002DBE8046c43FEfE86ec78b1112247BB8";
-    const bbAmUsdt = "0xFf4ce5AAAb5a627bf82f4A571AB1cE94Aa365eA6";
-    const splitterAddress = await vault.splitter();
-    const insurance = await vault.insurance();
-    const block = await hre.ethers.provider.getBlock("latest");
-
-    // console.log("!", user.address, signer.address, balancerPool, bbAmUsdc);
-    // // await IERC20__factory.connect(bbAmUsdc, user).balanceOf(balancerPool);
-    // console.log("!!", user.address, signer.address, balancerPool, bbAmUsdt);
-    // await IERC20__factory.connect(bbAmUsdt, signer).balanceOf(balancerPool);
-    // console.log("!!!");
-    // await IERC20__factory.connect(bbAmDai, user).balanceOf(balancerPool);
-    console.log("!!!!");
-
-    const dest = {
-      title: title || "no-name",
-      block: block.number,
-      blockTimestamp: block.timestamp,
-      signer: {
-        usdc: await IERC20__factory.connect(MaticAddresses.USDC_TOKEN, user).balanceOf(signer.address),
-      },
-      user: {
-        usdc: await IERC20__factory.connect(MaticAddresses.USDC_TOKEN, user).balanceOf(user.address),
-      },
-      strategy: {
-        usdc: await IERC20__factory.connect(MaticAddresses.USDC_TOKEN, user).balanceOf(strategy.address),
-        usdt: await IERC20__factory.connect(MaticAddresses.USDT_TOKEN, user).balanceOf(strategy.address),
-        dai: await IERC20__factory.connect(MaticAddresses.DAI_TOKEN, user).balanceOf(strategy.address),
-        bal: await IERC20__factory.connect(MaticAddresses.BAL_TOKEN, user).balanceOf(strategy.address),
-        bptPool: await IERC20__factory.connect(balancerPool, user).balanceOf(strategy.address),
-        totalAssets: await strategy.totalAssets(),
-        investedAssets: await StrategyBaseV2__factory.connect(strategy.address, user).investedAssets()
-      },
-      gauge: {
-        strategyBalance: await IBalancerGauge__factory.connect(gauge, user).balanceOf(strategy.address),
-      },
-      balancerPool: {
-        bbAmUsdc: await IERC20__factory.connect(bbAmUsdc, user).balanceOf(balancerPool),
-        bbAmUsdt: await IERC20__factory.connect(bbAmUsdt, user).balanceOf(balancerPool),
-        bbAmDai: await IERC20__factory.connect(bbAmDai, user).balanceOf(balancerPool),
-      },
-      splitter: {
-        usdc: await IERC20__factory.connect(MaticAddresses.USDC_TOKEN, user).balanceOf(splitterAddress),
-        totalAssets: await ISplitter__factory.connect(splitterAddress, user).totalAssets(),
-      },
-      vault: {
-        usdc: await IERC20__factory.connect(MaticAddresses.USDC_TOKEN, user).balanceOf(vault.address),
-        userShares: await vault.balanceOf(user.address),
-        signerShares: await vault.balanceOf(signer.address),
-        userUsdc: await vault.convertToAssets(await vault.balanceOf(user.address)),
-        signerUsdc: await vault.convertToAssets(await vault.balanceOf(signer.address)),
-        sharePrice: await vault.sharePrice(),
-        totalSupply: await vault.totalSupply(),
-        totalAssets: await vault.totalAssets(),
-      },
-      insurance: {
-        usdc: await IERC20__factory.connect(MaticAddresses.USDC_TOKEN, user).balanceOf(insurance),
-      },
-    }
-
-
-    console.log("State", dest);
-    return dest;
-  }
-
-  static async saveListStatesToCSV(pathOut: string, states: IState[]) {
-
-    const headers = [
-      "title",
-      "block",
-      "timestamp",
-
-      "$signer",
-      "$user",
-
-      "vault-user-shares",
-      "vault-signer-shares",
-
-      "vault-$user",
-      "vault-signer",
-
-      "sharePrice-vault",
-      "totalSupply-vault",
-      "totalAssets-vault",
-
-      "$insurance",
-      "$strategy",
-      "usdt-strategy",
-      "dai-strategy",
-      "bal-strategy",
-      "bptp-strategy",
-      "totalAssets-strategy",
-      "investedAssets-strategy",
-      "bptp-gauge",
-      "$vault",
-      "$splitter",
-      "totalAssets-splitter",
-      "bbAmUsdc-pool",
-      "bbAmUsdt-pool",
-      "bbAmDai-pool",
-    ];
-    const decimalsSharedPrice = 6;
-    const decimalsUSDC = 6;
-    const decimalsUSDT = 6;
-    const decimalsDAI = 18;
-    const decimalsBAL = 18;
-    const decimalsBbAmUsdc = 18;
-    const decimalsBbAmUsdt = 18;
-    const decimalsBbAmDai = 18;
-    const decimalsBptp = 18;
-    const decimals = [
-      0,
-      0,
-      0,
-      decimalsUSDC, // signer.usdc
-      decimalsUSDC, // user.usdc
-      decimalsUSDC, // vault.userShares
-      decimalsUSDC, // vault.signerShares
-      decimalsUSDC, // vault.userUsdc
-      decimalsUSDC, // vault.signerUsdc
-      decimalsSharedPrice, // shared price
-      decimalsUSDC, // vault.totlaSupply
-      decimalsUSDC, // vault.totalAssets
-      decimalsUSDC, // insurance.usdc
-      decimalsUSDC, // strategy.usdc
-      decimalsUSDT, // strategy.usdt
-      decimalsDAI, // strategy.dai
-      decimalsBAL, // strategy.bal
-      decimalsBptp, // strategy.bptPool
-      decimalsUSDC, // strategy.totalAssets
-      decimalsUSDC, // strategy.investedAssets
-      decimalsBptp, // gauge.strategyBalance
-      decimalsUSDC, // vault.usdc
-      decimalsUSDC, // splitter.usdc
-      decimalsUSDC, // splitter.totalAssets,
-      decimalsBbAmUsdc,
-      decimalsBbAmUsdt,
-      decimalsBbAmDai
-    ];
-    writeFileSync(pathOut, headers.join(";") + "\n", {encoding: 'utf8', flag: "a" });
-    for (const item of states) {
-      const line = [
-        item.title,
-        item.block,
-        item.blockTimestamp,
-
-        item.signer.usdc,
-        item.user.usdc,
-
-        item.vault.userShares,
-        item.vault.signerShares,
-
-        item.vault.userUsdc,
-        item.vault.signerUsdc,
-
-        item.vault.sharePrice,
-        item.vault.totalSupply,
-        item.vault.totalAssets,
-
-        item.insurance.usdc,
-        item.strategy.usdc,
-        item.strategy.usdt,
-        item.strategy.dai,
-        item.strategy.bal,
-        item.strategy.bptPool,
-        item.strategy.totalAssets,
-        item.strategy.investedAssets,
-        item.gauge.strategyBalance,
-        item.vault.usdc,
-        item.splitter.usdc,
-        item.splitter.totalAssets,
-        item.balancerPool.bbAmUsdc,
-        item.balancerPool.bbAmUsdt,
-        item.balancerPool.bbAmDai
-      ];
-      writeFileSync(pathOut,
-        line.map((x, index) =>
-          typeof x === "object"
-            ? +formatUnits(x, decimals[index])
-            : "" + x
-        ).join(";") + "\n",
-        {encoding: 'utf8', flag: "a"}
-      );
-    }
-  }
-
-  static getTotalUsdAmount(state: IState) : BigNumber {
-    return state.user.usdc.add(
-      state.signer.usdc
-    ).add(
-      state.vault.usdc
-    ).add(
-      state.insurance.usdc
-    ).add(
-      state.strategy.usdc
-    ).add(
-      state.splitter.usdc
-    );
-  }
-
-  /**
-   * Get initial state marked as "enter"
-   * and final state marked as "final"
-   * @param states
-   */
-  static outputProfit(states: IState[]) {
-    if (states.length < 2) return;
-
-    const enter: IState = states[0];
-    const final: IState = states[states.length - 1];
-
-    // ethereum timestamp is in seconds
-    // https://ethereum.stackexchange.com/questions/7853/is-the-block-timestamp-value-in-solidity-seconds-or-milliseconds
-    const timeSeconds = (final.blockTimestamp - enter.blockTimestamp);
-    const initialAmount = this.getTotalUsdAmount((enter));
-    const finalAmount = this.getTotalUsdAmount(final);
-    const amount = finalAmount.sub(initialAmount);
-    const amountNum = +formatUnits(amount, 6);
-    const apr = amountNum * 365
-      / (timeSeconds / (24*60*60))
-      / +formatUnits(initialAmount, 6)
-      * 100;
-    console.log("final.blockTimestamp", final.blockTimestamp);
-    console.log("enter.blockTimestamp", enter.blockTimestamp);
-    console.log("final.getTotalUsdAmount", this.getTotalUsdAmount(final));
-    console.log("final.getTotalUsdAmount", this.getTotalUsdAmount(enter));
-    console.log("Initial amount", initialAmount);
-    console.log("Final amount", initialAmount);
-    console.log("Total profit", amountNum);
-    console.log("Duration in seconds", timeSeconds);
-    console.log("Duration in days", timeSeconds / (24*60*60));
-    console.log("Estimated APR, %", apr);
-
-  }
-
   static async makeStrategyDeployer(
     core: CoreAddresses,
     asset: string,
@@ -382,8 +98,7 @@ export class UniversalTestUtils {
   /**
    * Finds event "LossCovered(lossValue)" in {tx}, returns {lossValue}
    */
-  static async extractLossCovered(tx: ContractTransaction, vaultAddress: string) : Promise<BigNumber | undefined> {
-    const cr = await tx.wait();
+  static async extractLossCovered(cr: ContractReceipt, vaultAddress: string) : Promise<BigNumber | undefined> {
     if (cr.events) {
       for (const event of cr.events) {
         if (event.address === vaultAddress) {
@@ -397,5 +112,39 @@ export class UniversalTestUtils {
         }
       }
     }
+  }
+
+  static async extractDistributed(cr: ContractReceipt, forwarder: string) : Promise<IDistributedInfo[]> {
+    const dest: IDistributedInfo[] = [];
+    if (cr.events) {
+      for (const event of cr.events) {
+        if (event.address === forwarder) {
+          if (event.event === "Distributed") {
+            if (event.args) {
+              console.log("Distributed", event.args);
+              dest.push({
+                sender: event.args[0],
+                incomeToken: event.args[1],
+                queuedBalance: event.args[2],
+                tetuValue: event.args[3],
+                tetuBalance: event.args[4],
+                toInvestFund: event.args[5],
+                toGauges: event.args[6],
+                toBribes: event.args[7]
+              });
+            }
+          }
+        }
+      }
+    }
+
+    return dest;
+  }
+
+  static async getAnOperator(strategy: string, signer: SignerWithAddress) : Promise<SignerWithAddress> {
+    const controller = await StrategyBaseV2__factory.connect(strategy, signer).controller();
+    const controllerAsUser = await ControllerV2__factory.connect(controller, signer);
+    const operators = await controllerAsUser.operatorsList();
+    return Misc.impersonate(operators[0]);
   }
 }
